Calc of Z points from Cycloid parameter t(^Z)
Calculate Z values of the t^Z (where t^1 is one of the parameter of the standard cycloid function) in the area t 0..1. For the reason ellipse perimeter calculation is a modified shortened cycloid function, the factor Z (instead of 1) has to be calculated for t in the area 0..1. By calculating points in the area 0..1, a curve fitting exercise will start. An achieved result so far: (intermediate result) Z(t) = ((2.58492 * t + 1)**(-1.29308))*exp(t)*(-0.1*t+0.00885128)*ln(1.56457*t)+1.153. More to come